Fixed converting unit types
Top appends the unit type in the RES,VIRT and SHR columns it's not consistent
so VIRT can go into into 1.5g while still displaying m but RES converts
as soon as it hits 1.0g. The previous version also didn't account for kb not
being added to basic readouts and incorrectly presented it as m on the graphs.